requirements.txt
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
anthropic==0.7.8
openai==1.3.7
google-api-python-client==2.108.0
airtable-python-wrapper==0.15.3
python-dotenv==1.0.0
requests==2.31.0
asyncio
json
datetime
typing
pathlib
os
sys
